Add WIP generate_cf_package.py command to compile CF releases in to a single package for testing
